Files
From file creation to news dissemination.
A little modification is required. In it if a news is posted then the
broker reads the news and creates a text file in the same folder. The
subscriber has to read the file according to its interest. This all
works correctly but the catch is time filtering of news is not done.